requirements.txt
langfuse>=3.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
